requirements.txt
flask
google-genai
python-dotenv
pydantic==2.8.0
gunicorn==21.2.0

# Evaluation dependencies
requests>=2.31.0
statistics-python>=0.3.0